+++ packages/astro/e2e/fixtures/dev-toolbar/astro.config.mjs
@@ -1,5 +1,6 @@
 import preact from '@astrojs/preact';
+import { myIntegration } from './custom-integration.js';
 
 export default {
-	integrations: [preact()],
+	integrations: [preact(), myIntegration()],
 };