+++ packages/integrations/node/test/prerender.test.js
@@ -82,8 +82,7 @@ describe('Prerendering', () => {
 			});
 			const html = await res.text();
 			const $ = cheerio.load(html);
-			assert.equal(res.status, 301);
-			assert.equal(res.headers.get('location'), '/some-base/two/');
+			assert.equal(res.status, 200);
 			assert.equal($('h1').text(), 'Two');
 		});
 	});
@@ -171,8 +170,8 @@ describe('Prerendering', () => {
 			const html = await res.text();
 			const $ = cheerio.load(html);
 
-			expect(res.status).to.equal(200);
-			expect($('h1').text()).to.equal('One');
+			assert.equal(res.status, 200);
+			assert.equal($('h1').text(), 'One');
 		});
 
 		it('Can render prerendered route', async () => {
@@ -180,8 +179,8 @@ describe('Prerendering', () => {
 			const html = await res.text();
 			const $ = cheerio.load(html);
 
-			expect(res.status).to.equal(200);
-			expect($('h1').text()).to.equal('Two');
+			assert.equal(res.status, 200);
+			assert.equal($('h1').text(), 'Two');
 		});
 	});
 });
@@ -201,7 +200,7 @@ describe('Hybrid rendering', () => {
 				adapter: nodejs({ mode: 'standalone' }),
 			});
 			await fixture.build();
-			const { startServer } = await await load();
+			const { startServer } = await load();
 			let res = startServer();
 			server = res.server;
 			await waitServerListen(server.server);
@@ -252,8 +251,9 @@ describe('Hybrid rendering', () => {
 			const res = await fetch(`http://${server.host}:${server.port}/some-base/one`, {
 				redirect: 'manual',
 			});
-			assert.equal(res.status, 301);
-			assert.equal(res.headers.get('location'), '/some-base/one/');
+			const html = await res.text();
+			const $ = cheerio.load(html);
+			assert.equal(res.status, 200);
 			assert.equal($('h1').text(), 'One');
 		});
 	});
@@ -267,7 +267,7 @@ describe('Hybrid rendering', () => {
 				adapter: nodejs({ mode: 'standalone' }),
 			});
 			await fixture.build();
-			const { startServer } = await await load();
+			const { startServer } = await load();
 			let res = startServer();
 			server = res.server;
 			await waitServerListen(server.server);
@@ -313,6 +313,37 @@ describe('Hybrid rendering', () => {
 
 			assert.equal(res.status, 200);
 			assert.equal($('h1').text(), 'One');
+		});
+	});
+
+	describe('Shared modules', async () => {
+		before(async () => {
+			process.env.PRERENDER = false;
+
+			fixture = await loadFixture({
+				root: './fixtures/prerender/',
+				output: 'hybrid',
+				adapter: nodejs({ mode: 'standalone' }),
+			});
+			await fixture.build();
+			const { startServer } = await load();
+			let res = startServer();
+			server = res.server;
+		});
+
+		after(async () => {
+			await server.stop();
+			await fixture.clean();
+			delete process.env.PRERENDER;
+		});
+
+		it('Can render SSR route', async () => {
+			const res = await fetch(`http://${server.host}:${server.port}/third`);
+			const html = await res.text();
+			const $ = cheerio.load(html);
+
+			assert.equal(res.status, 200);
+			assert.equal($('h1').text(), 'shared');
 		});
 	});
 });